pyproject.toml
[tool.poetry]
name = "smbscan"
version = "0.1.0"
description = "tool developed to enumerate file shares on an internal network."
authors = ["P"]
license = "GNU GENERAL PUBLIC LICENSE"

[tool.poetry.dependencies]
python = "^3.10"
impacket = "^0.10.0"
python-slugify = "^6.1.2"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"

[tool.pytest.ini_options]
addopts = "-ra -q --cov"

[tool.isort]
profile = "black"
line_length = 80


[tool.black]
line-length = 80

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
